// apps/web-svelte/src/lib/api/generated/model/fileArchivePreparation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface FileArchivePreparation {
  id: string;
  fileName: string;
  ready: boolean;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  progressPercent: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  processedFiles: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalFiles: number | string;
  /** @nullable */
  error: string | null;
}